Initial implementation for the JVM and JavaScript.
The Scala Native version is not there yet, as Scala Native does not
have yet the necessary mechanisms to implement the API.

The only discrepancy between JVM and JS is the behavior of
`protected` constructors. The JVM version considers that they are
visible, because they are compiled to `public` in the bytecode,
while Scala.js explicitly excludes them. It seems that the most
reasonable fix would be to allow access to protected constructors
in Scala.js' core `Reflect` mechanism.