PROJECT_NAME handling is essential, despite misleading log file entry.
The variable PROJECT_NAME shows up in the log file as unsubstituted text
PROJECT_NAME. That text is apparently unused in the .yml script.
However, PROJECT_NAME appears to be substituted as it is passed to
sbt. The substituted value is then used in build.sbt to distinguish
between JVM, Scala.js, and Scala Native sub-projects.  Those sub-projects
do not appear as such in the build.sbt file, but are supplied behind
the scene by CrossProject.

It makes no sense to spend further time trying to get the substitution
for PROJECT_NAME to display in the log file, since the .yml will
soon have to be ported to another CI, which will probably have different
substitution rules.